S3移行のための修正

+++ app/models/exhibit_information.rb
@@ -21,20 +21,43 @@ class ExhibitInformation < ApplicationRecord
 
   scope :order_by_title_in_event, -> (event) { where(event_id: event.id).order(:title) }
 
+  def image_url
+    if Rails.env.development?
+      return ActionController::Base.helpers.asset_path('noimage.jpg') if image.blank?
+      image
+    else
+      # 本番環境ではS3の画像を参照する
+      return ActionController::Base.helpers.asset_path('noimage.jpg') if image.blank?
+      S3Facade.new.get_object_url(Rails.application.config.s3_url, image)
+    end
+  end
+
   def copy_image(source_image_path)
     file_name = File.basename(source_image_path)
     dest_suffix = 'uploads/exhibit_informations/image/' + id.to_s
-    dest_dir = Rails.root.join(Rails.application.config.exhibit_informations_image_root, dest_suffix)
-    dest_full_path = dest_dir.join(file_name)
-    # image_tagでupload以下を呼び出す際は先頭に/をつける必要がある
-    image_tag_path = '/' + dest_suffix + '/' + file_name
-    if File.exist?(dest_dir)
-      FileUtils.remove(dest_dir.glob('*'))
+    
+    if Rails.env.development?
+      dest_dir = Rails.root.join(Rails.application.config.exhibit_informations_image_root, dest_suffix)
+      dest_full_path = dest_dir.join(file_name)
+      # image_tagでupload以下を呼び出す際は先頭に/をつける必要がある
+      image_tag_path = '/' + dest_suffix + '/' + file_name
+      if File.exist?(dest_dir)
+        FileUtils.remove(dest_dir.glob('*'))
+      else
+        FileUtils.mkdir_p(dest_dir)
+      end
+      FileUtils.cp(source_image_path, dest_full_path)
+      self.image = image_tag_path
     else
-      FileUtils.mkdir_p(dest_dir)
+      file_path = dest_suffix + '/' + file_name
+      begin
+        S3Facade.new.copy_object(source_image_path, file_path)
+      rescue Aws::S3::Errors::ServiceError => e
+        Rails.logger.error("Failed to copy object from S3: #{e.message}")
+      end
+      # S3移行前のURLと同じにするために先頭にスラッシュをつける
+      self.image = '/' + file_path
     end
-    FileUtils.cp(source_image_path, dest_full_path)
-    self.image = image_tag_path
   end
 
   def youtube_widget_url
